add db.js
